Add the Android Application project.
The game is compiled as a dynamic library by the project in android/lib.

The Java part of the game is in android/java. It loads the library and calls a
method from it.

The build script has been modified a lot to finally compile and being able to
launch the application on the phone with no error. These modifications are as
follows:

libclaw — We use the full paths to the libraries, otherwise the compilation uses
the ones of the system.

SDL — The library has been compiled as a static library. The same goes for the
mixer part.

Install prefix — The install prefix is android/java, from asgp's source root.
The data is installed in android/java/assets in order to be able to load the
resources by their path. The library is installed in android/java/libs

I had to copy gnustl_shared in android/java/libs too, otherwise there was some
missing symbols at run time.