Better finger handling.
level_ending_effect is ready to process the finger events to skip the scores.

The cart jumps or crouch if the finger slides up to down or down to up.

The threshold to throw a plunger is a slide shorter than 80 units. A cannonball
is thrown otherwise, at the initial position of the finger.