Android: the back button triggers the home screen.
In Android builds, pressing the back button of the phone when in the level
selection screen will react like the home button (i.e. the game goes to the
background).

The button is handled by the new class back_home_button_item, which calls the
right method from the Java class.